+++ exporting/tests/test_exporting_engine.c
@@ -1,369 +1,14 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "libnetdata/libnetdata.h"
+#include "test_exporting_engine.h"
 #include "libnetdata/required_dummies.h"
-
-#include "exporting/exporting_engine.h"
-#include "exporting/graphite/graphite.h"
-
-#include <stdarg.h>
-#include <stddef.h>
-#include <setjmp.h>
-#include <stdint.h>
-#include <cmocka.h>
 
 RRDHOST *localhost;
 netdata_rwlock_t rrd_rwlock;
 
-#define MAX_LOG_LINE 1024
 char log_line[MAX_LOG_LINE + 1];
 
-// Use memomy allocation functions guarded by CMocka in strdupz
-const char *__wrap_strdupz(const char *s)
-{
-    char *duplicate = malloc(sizeof(char) * (strlen(s) + 1));
-    strcpy(duplicate, s);
-
-    return duplicate;
-}
-
-time_t __wrap_now_realtime_sec(void)
-{
-    function_called();
-    return mock_type(time_t);
-}
-
-void __wrap_info_int(const char *file, const char *function, const unsigned long line, const char *fmt, ...)
-{
-    (void)file;
-    (void)function;
-    (void)line;
-
-    function_called();
-
-    va_list args;
-
-    va_start(args, fmt);
-    vsnprintf(log_line, MAX_LOG_LINE, fmt, args);
-    va_end(args);
-}
-
-struct engine *__real_read_exporting_config();
-struct engine *__wrap_read_exporting_config()
-{
-    function_called();
-    return mock_ptr_type(struct engine *);
-}
-
-static struct engine *__mock_read_exporting_config()
-{
-    struct engine *engine = calloc(1, sizeof(struct engine));
-    engine->config.prefix = strdupz("netdata");
-    engine->config.hostname = strdupz("test-host");
-    engine->config.update_every = 3;
-    engine->config.options = BACKEND_SOURCE_DATA_AVERAGE | BACKEND_OPTION_SEND_NAMES;
-
-    engine->connector_root = calloc(1, sizeof(struct connector));
-    engine->connector_root->config.type = BACKEND_TYPE_GRAPHITE;
-    engine->connector_root->engine = engine;
-
-    engine->connector_root->instance_root = calloc(1, sizeof(struct instance));
-    struct instance *instance = engine->connector_root->instance_root;
-    instance->connector = engine->connector_root;
-    instance->config.destination = strdupz("localhost");
-    instance->config.update_every = 1;
-    instance->config.buffer_on_failures = 10;
-    instance->config.timeoutms = 10000;
-    instance->config.charts_pattern = strdupz("*");
-    instance->config.hosts_pattern = strdupz("localhost *");
-    instance->config.send_names_instead_of_ids = 1;
-
-    return engine;
-}
-
-int __real_init_connectors(struct engine *engine);
-int __wrap_init_connectors(struct engine *engine)
-{
-    function_called();
-    check_expected_ptr(engine);
-    return mock_type(int);
-}
-
-int __wrap_mark_scheduled_instances(struct engine *engine)
-{
-    function_called();
-    check_expected_ptr(engine);
-    return mock_type(int);
-}
-
-int __real_prepare_buffers(struct engine *engine);
-int __wrap_prepare_buffers(struct engine *engine)
-{
-    function_called();
-    check_expected_ptr(engine);
-    return mock_type(int);
-}
-
-int __wrap_notify_workers(struct engine *engine)
-{
-    function_called();
-    check_expected_ptr(engine);
-    return mock_type(int);
-}
-
-int __wrap_send_internal_metrics(struct engine *engine)
-{
-    function_called();
-    check_expected_ptr(engine);
-    return mock_type(int);
-}
-
-int __mock_start_batch_formatting(struct instance *instance)
-{
-    function_called();
-    check_expected_ptr(instance);
-    return mock_type(int);
-}
-
-int __mock_start_host_formatting(struct instance *instance)
-{
-    function_called();
-    check_expected_ptr(instance);
-    return mock_type(int);
-}
-
-int __mock_start_chart_formatting(struct instance *instance)
-{
-    function_called();
-    check_expected_ptr(instance);
-    return mock_type(int);
-}
-
-int __mock_metric_formatting(struct instance *instance, RRDDIM *rd)
-{
-    function_called();
-    check_expected_ptr(instance);
-    check_expected_ptr(rd);
-    return mock_type(int);
-}
-
-int __mock_end_chart_formatting(struct instance *instance)
-{
-    function_called();
-    check_expected_ptr(instance);
-    return mock_type(int);
-}
-
-int __mock_end_host_formatting(struct instance *instance)
-{
-    function_called();
-    check_expected_ptr(instance);
-    return mock_type(int);
-}
-
-int __mock_end_batch_formatting(struct instance *instance)
-{
-    function_called();
-    check_expected_ptr(instance);
-    return mock_type(int);
-}
-
-void __rrdhost_check_rdlock(RRDHOST *host, const char *file, const char *function, const unsigned long line)
-{
-    (void)host;
-    (void)file;
-    (void)function;
-    (void)line;
-}
-
-void __rrdset_check_rdlock(RRDSET *st, const char *file, const char *function, const unsigned long line)
-{
-    (void)st;
-    (void)file;
-    (void)function;
-    (void)line;
-}
-
-void __rrd_check_rdlock(const char *file, const char *function, const unsigned long line)
-{
-    (void)file;
-    (void)function;
-    (void)line;
-}
-
-void __wrap_uv_thread_create(uv_thread_t thread, void (*worker)(void *arg), void *arg)
-{
-    function_called();
-
-    check_expected_ptr(thread);
-    check_expected_ptr(worker);
-    check_expected_ptr(arg);
-}
-
-void __wrap_uv_mutex_lock(uv_mutex_t *mutex)
-{
-    (void)mutex;
-}
-
-void __wrap_uv_mutex_unlock(uv_mutex_t *mutex)
-{
-    (void)mutex;
-}
-
-void __wrap_uv_cond_signal(uv_cond_t *cond_var)
-{
-    (void)cond_var;
-}
-
-void __wrap_uv_cond_wait(uv_cond_t *cond_var, uv_mutex_t *mutex)
-{
-    (void)cond_var;
-    (void)mutex;
-}
-
-ssize_t __wrap_recv(int sockfd, void *buf, size_t len, int flags)
-{
-    function_called();
-
-    check_expected(sockfd);
-    check_expected_ptr(buf);
-    check_expected(len);
-    check_expected(flags);
-
-    char *mock_string = "Test recv";
-    strcpy(buf, mock_string);
-
-    return strlen(mock_string);
-}
-
-ssize_t __wrap_send(int sockfd, const void *buf, size_t len, int flags)
-{
-    function_called();
-
-    check_expected(sockfd);
-    check_expected_ptr(buf);
-    check_expected_ptr(buf);
-    check_expected(len);
-    check_expected(flags);
-
-    return strlen(buf);
-}
-
-int __wrap_connect_to_one_of(
-    const char *destination,
-    int default_port,
-    struct timeval *timeout,
-    size_t *reconnects_counter,
-    char *connected_to,
-    size_t connected_to_size)
-{
-    (void)timeout;
-
-    function_called();
-
-    check_expected(destination);
-    check_expected_ptr(default_port);
-    // TODO: check_expected_ptr(timeout);
-    check_expected(reconnects_counter);
-    check_expected(connected_to);
-    check_expected(connected_to_size);
-
-    return mock_type(int);
-}
-
-static int setup_configured_engine(void **state)
-{
-    struct engine *engine = __mock_read_exporting_config();
-
-    engine->after = 1;
-    engine->before = 2;
-
-    *state = engine;
-
-    return 0;
-}
-
-static int teardown_configured_engine(void **state)
-{
-    struct engine *engine = *state;
-
-    struct instance *instance = engine->connector_root->instance_root;
-    free((void *)instance->config.destination);
-    free(instance->config.charts_pattern);
-    free(instance->config.hosts_pattern);
-    free(instance);
-
-    free(engine->connector_root);
-
-    free((void *)engine->config.prefix);
-    free((void *)engine->config.hostname);
-    free(engine);
-
-    return 0;
-}
-
-static int setup_rrdhost()
-{
-    localhost = calloc(1, sizeof(RRDHOST));
-    localhost->tags = strdupz("TAG1=VALUE1 TAG2=VALUE2");
-
-    localhost->rrdset_root = calloc(1, sizeof(RRDSET));
-    RRDSET *st = localhost->rrdset_root;
-    st->rrdhost = localhost;
-    strcpy(st->id, "chart_id");
-    st->name = strdupz("chart_name");
-
-    localhost->rrdset_root->dimensions = calloc(1, sizeof(RRDDIM));
-    RRDDIM *rd = localhost->rrdset_root->dimensions;
-    rd->rrdset = st;
-    rd->id = strdupz("dimension_id");
-    rd->name = strdupz("dimension_name");
-    rd->last_collected_value = 123000321;
-    rd->last_collected_time.tv_sec = 15051;
-    rd->next = NULL;
-
-    return 0;
-}
-
-static int teardown_rrdhost()
-{
-    RRDDIM *rd = localhost->rrdset_root->dimensions;
-    free((void *)rd->name);
-    free((void *)rd->id);
-    free(rd);
-
-    RRDSET *st = localhost->rrdset_root;
-    free((void *)st->name);
-    free(st);
-
-    free((void *)localhost->tags);
-    free(localhost);
-
-    return 0;
-}
-
-static void init_connectors_in_tests(struct engine *engine);
-static int setup_initialized_engine(void **state)
-{
-    setup_configured_engine(state);
-
-    struct engine *engine = *state;
-    init_connectors_in_tests(engine);
-
-    setup_rrdhost();
-
-    return 0;
-}
-
-static int teardown_initialized_engine(void **state)
-{
-    teardown_rrdhost();
-    teardown_configured_engine(state);
-
-    return 0;
-}
-
-static void init_connectors_in_tests(struct engine *engine)
+void init_connectors_in_tests(struct engine *engine)
 {
     expect_function_call(__wrap_uv_thread_create);
